Experience, Damage and Display Tweaks + Barbarian Fix
- Tweaked config values regarding XP per Level.
- Brought Damage values on Weapons back in line with the weapon's material type.
- Fixed Barbarians not doing damage to players.

+++ src/main/java/com/thexfactor117/lsc/capabilities/cap/CapabilityLSCPlayer.java
@@ -74,7 +74,7 @@ public class CapabilityLSCPlayer
 				nbt.setInteger("Mana", instance.getMana());
 				nbt.setInteger("ManaPerSecond", instance.getManaPerSecond());
 				
-				nbt.setInteger("HealthPerSecond", instance.getHealthPerSecond());
+				nbt.setDouble("HealthPerSecond", instance.getHealthPerSecond());
 				
 				nbt.setDouble("CriticalChance", instance.getCriticalChance());
 				nbt.setDouble("CriticalDamage", instance.getCriticalDamage());
